requirements.txt
RPi.GPIO==0.7.1
spidev==3.6
textwrap3==0.9.2
python-dotenv==1.0.1
openai==1.12.0
requests~=2.28.1
pylint~=3.1.0